# specs/001-ai-client-documentation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Warba Bank — AI-Powered Client Documentation API
  version: 1.0.0
  description: |
    Backend contract for Track 1 of the Warba Bank Corporate Banking AI Challenge.

    Governed by Constitution v1.0.0. Three contract-level invariants are worth stating
    before the endpoints, because they are the reason several endpoints look the way
    they do:

    1. **No auto-approval path exists.** `POST /documents/{id}/approve` is the only
       transition into `APPROVED`, it requires an authenticated RM, and it requires the
       caller to echo the `content_hash` of the exact version being approved. There is no
       endpoint, parameter, or scheduled process that approves a document any other way
       (Principle III, FR-026/FR-027).
    2. **Failure produces no document.** Generation, screening, and retrieval failures
       return a 4xx/5xx error, never a partial draft (Principle I, FR-037).
    3. **Audit is read-only over HTTP.** No endpoint mutates or deletes an audit event.
       The database role backing this API does not hold UPDATE or DELETE on the audit
       table (Principle VIII, FR-032).

    All data served by this API is synthetic (Principle VII).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /documents/{document_id}/approve:
    post:
      tags: [Approval]
      summary: Approve a document — the only path to APPROVED
      description: |
        The single transition into `APPROVED`, and the constitutional heart of this API
        (Principle III).

        Every one of these must hold, or the request is refused:

        - the caller's role is `RM` **and** they own the client's portfolio (403 otherwise);
        - `content_hash` matches the current version exactly — the RM approves a specific
          version, not "the document" (412 otherwise);
        - zero unresolved gaps remain across all sections (422 otherwise, FR-025);
        - the current version passed deterministic screening (451 otherwise);
        - the document is not already approved (409 otherwise).

        `acknowledge_gaps` lets the RM accept a gap deliberately, with a note. It resolves the
        gap — it does not bypass the check. The acknowledgement is recorded in the approval.

        There is no timeout, default, inactivity, or administrative path to this state
        (FR-027).
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_hash, confirm_reviewed]
              properties:
                content_hash:
                  type: string
                  description: Hash of the exact version being approved
                confirm_reviewed:
                  type: boolean
                  enum: [true]
                  description: |
                    Explicit affirmation that the RM has reviewed the content and accepts
                    authorship. Must be literally true — a deliberate act, not a default.
                acknowledge_gaps:
                  type: array
                  items:
                    type: object
                    required: [section_key, field, note]
                    properties:
                      section_key: { type: string }
                      field: { type: string }
                      note: { type: string }
      responses:
        '200':
          description: Approved; ApprovalRecord and audit event written
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovalRecord' }
        '403':
          description: Caller is not an RM, or does not own this client's portfolio
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Already approved (terminal state)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '412':
          description: content_hash does not match the current version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Unresolved gaps remain (FR-025)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnresolvedGapsError' }
        '451':
          description: Current version has not passed Shariah screening
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScreeningBlockedError' }

  # ──────────────────────────── Export ────────────────────────────
...
  /audit/events:
    get:
      tags: [Audit]
      summary: Query the audit trail
      description: |
        Read-only. Requires `COMPLIANCE` role (FR-034). There is no create, update, or delete
        counterpart anywhere in this API — the backing database role holds INSERT and SELECT
        only (FR-032, R6).
      parameters:
        - { name: document_id, in: query, schema: { type: string, format: uuid } }
        - { name: client_reference, in: query, schema: { type: string } }
        - { name: actor_id, in: query, schema: { type: string, format: uuid } }
        - { name: event_type, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 1000 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEvent' }
                  total: { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
  /audit/documents/{document_id}/lifecycle:
    get:
      tags: [Audit]
      summary: Full reconstructable lifecycle of one document
      description: |
        The Compliance reconstruction view (US4). Returns every generation, edit,
        regeneration, rejection, screening block, and approval in order, with the inputs,
        model, template, and prompt versions used at each step — enough to answer "how was
        this produced" without reading the code (Principle VIII, SC-013).
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentLifecycle' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /audit/export:
    get:
      tags: [Audit]
      summary: Export audit records in a machine-readable format
      description: Satisfies FR-035. The export is itself recorded as an `AUDIT_EXPORTED` event.
      parameters:
        - { name: document_id, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - name: format
          in: query
          schema: { type: string, enum: [json, csv], default: json }
      responses:
        '200':
          description: Exported records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEvent' }
            text/csv:
              schema: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Forbidden:
      description: Authenticated but not permitted for this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    DocumentType:
      type: string
      enum: [CALL_REPORT, CLIENT_PROFILE, CREDIT_MEMO_NARRATIVE, KYC_SUMMARY]
    DocumentStatus:
      type: string
      enum: [DRAFT, UNDER_REVIEW, REJECTED, APPROVED]
    ShariahStatus:
      type: string
      enum: [PENDING_REVIEW, CLEARED, FLAGGED]
      description: |
        Defaults to PENDING_REVIEW and is never set to CLEARED by the system. The system
        prepares documents for Shariah review; it does not clear them (Principle II).
...
    DocumentSummary:
      type: object
      required: [id, client_id, document_type, status, shariah_status, created_at]
      properties:
        id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        client_reference: { type: string }
        document_type: { $ref: '#/components/schemas/DocumentType' }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        shariah_status: { $ref: '#/components/schemas/ShariahStatus' }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
...
    DocumentVersionSummary:
      type: object
      required: [id, version_number, origin, created_by, content_hash, created_at]
      properties:
        id: { type: string, format: uuid }
        version_number: { type: integer }
        origin: { type: string, enum: [GENERATED, REGENERATED_SECTION, RM_EDITED] }
        created_by: { type: string, format: uuid }
        created_by_name: { type: string }
        content_hash: { type: string }
        model_id: { type: string, nullable: true }
        template_version: { type: string }
        prompt_version: { type: string }
        created_at: { type: string, format: date-time }
    ApprovalRecord:
      type: object
      required: [id, document_id, version_id, approved_by, approver_name, approver_role, content_hash, approved_at]
      properties:
        id: { type: string, format: uuid }
        document_id: { type: string, format: uuid }
        version_id: { type: string, format: uuid }
        approved_by: { type: string, format: uuid }
        approver_name:
          type: string
          description: Snapshotted, so the record survives changes to the user account.
        approver_role: { type: string, enum: [RM] }
        content_hash: { type: string }
        shariah_status_at_approval: { $ref: '#/components/schemas/ShariahStatus' }
        gaps_acknowledged:
          type: array
          items:
            type: object
            properties:
              section_key: { type: string }
              field: { type: string }
              note: { type: string }
        approved_at: { type: string, format: date-time }
    AuditEvent:
      type: object
      required: [id, sequence, event_type, occurred_at, event_hash]
      properties:
        id: { type: string, format: uuid }
        sequence: { type: integer, format: int64 }
        event_type: { type: string }
        occurred_at: { type: string, format: date-time }
        actor_id: { type: string, format: uuid, nullable: true }
        actor_name: { type: string, nullable: true }
        client_reference: { type: string, nullable: true }
        document_id: { type: string, format: uuid, nullable: true }
        version_id: { type: string, format: uuid, nullable: true }
        document_type: { type: string, nullable: true }
        input_source_ids:
          type: array
          items: { type: string }
        model_id: { type: string, nullable: true }
        model_version: { type: string, nullable: true }
        prompt_version: { type: string, nullable: true }
        template_version: { type: string, nullable: true }
        output_hash: { type: string, nullable: true }
        detail:
          type: object
          description: |
            Identifiers and counts only. Never document content, client-bearing prompt text,
            or credentials (FR-042, NFR-SEC-04).
          additionalProperties: true
        prev_hash: { type: string, nullable: true }
        event_hash: { type: string }
    DocumentLifecycle:
      type: object
      required: [document, events, versions]
      properties:
        document: { $ref: '#/components/schemas/DocumentSummary' }
        versions:
          type: array
          items: { $ref: '#/components/schemas/DocumentVersionSummary' }
        approval:
          allOf: [{ $ref: '#/components/schemas/ApprovalRecord' }]
          nullable: true
        events:
          type: array
          items: { $ref: '#/components/schemas/AuditEvent' }
        chain_valid:
          type: boolean
          description: Hash chain integrity across this document's events.
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message:
          type: string
          description: |
            Plain, non-technical, and actionable — the RM must understand what happened and
            what to do next (FR-038, NFR-UX-02).
        detail: { type: object, additionalProperties: true }
    ScreeningBlockedError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            vocabulary_version: { type: string }
            findings:
              type: array
              items:
                type: object
                properties:
                  term: { type: string }
                  section_key: { type: string }
                  severity: { type: string }
                  rule_id: { type: string }
                  rationale: { type: string }
    UnresolvedGapsError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            unresolved_gaps:
              type: array
              items:
                type: object
                properties:
                  section_key: { type: string }
                  field: { type: string }
                  label: { type: string }